Explicit define govendor path during build process
When executing `make build` as a first time contributor I'm welcomed
with this error:

```
make build
>> fetching govendor
>> fetching dependencies
make: govendor: No such file or directory
make: *** [build] Error 1
```

This happens because we assume that `$GOPATH/bin` is available on
`$PATH` to locate the `govendor` binary.

Instead of instructing developers to configure their path, this commit
expands the assumption of the global go get intallation path into a
variable.

This would allow overriding of the location if necessary doing:

```
make build GOVENDOR=/usr/local/bin/govendor
```