Install dependencies before test step
As a first time contributor, when I've tried to execute the tests from a fresh
checkout I received the following error.

```
make test
 ==Running go test==
metrics/metrics.pb.go:20:8: cannot find package "github.com/gogo/protobuf/proto" in any of:
        /Users/bruno/go/src/github.com/digitalocean/do-agent/vendor/github.com/gogo/protobuf/proto (vendor tree)
        /usr/local/Cellar/go/1.8/libexec/src/github.com/gogo/protobuf/proto (from $GOROOT)
        /Users/bruno/go/src/github.com/gogo/protobuf/proto (from $GOPATH)
[...]
```

It happens because the dependencies are only installed when executing a
`make build`.

This commit adds a PHONY rule to install the dependencies, and wire then
up on both the `build` and `test` rules.