Adjust the README with instructions for first time Gophers
This commit adds instructions on how to checkout the code on your
computer and build.

This is needed to avoid friction with executing `got get` pointing to a
fork.

When a new contributor tries to `go get` from their fork, they will be
welcomed with an error message due to the import path still pointing to
the canonical project.

This commit adds the necessary information on how to checkout the
project and configure the remote as instructed by the community.