TS-1037
When the do-agent restarts, it is likely that it'll send a metric
faster than rate limitation will allow, resulting in a 429 error.
This can safely be ignored. The code adjustment herein determines
if the first response from a metric submission is an error with
a status of 429. If so, it ignores it.